+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # changelog
 
+### unreleased
+
+* Add VirtualMachine.WaitForNetIP helper
+
 ### 0.5.0 (2016-03-30)
 
 Generated fields using xsd type 'int' change to Go type 'int32'